refactor: user service 리팩터링
