+++ src/main/java/com/burntoburn/easyshift/controller/UserApiController.java
@@ -2,7 +2,7 @@ package com.burntoburn.easyshift.controller;
 
 import com.burntoburn.easyshift.dto.user.AddUserRequest;
 import com.burntoburn.easyshift.entity.user.User;
-import com.burntoburn.easyshift.service.UserService;
+import com.burntoburn.easyshift.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
